Segmented code and compartmentalized functionalities in different files
